+++ lib/command/setup_app.rb
@@ -6,6 +6,7 @@ module Command
     OPTIONS = [
       app_option(required: true),
       skip_secret_access_binding_option,
+      skip_secrets_setup_option,
       skip_post_creation_hook_option
     ].freeze
     DESCRIPTION = "Creates an app and all its workloads"
@@ -15,14 +16,16 @@ module Command
       - This should only be used for temporary apps like review apps, never for persistent apps like production or staging (to update workloads for those, use 'cpl apply-template' instead)
       - Configures app to have org-level secrets with default name "{APP_PREFIX}-secrets"
         using org-level policy with default name "{APP_PREFIX}-secrets-policy" (names can be customized, see docs)
-      - Use `--skip-secret-access-binding` to prevent the automatic setup of secrets
+      - Creates identity for secrets if it does not exist
+      - Use `--skip-secrets-setup` to prevent the automatic setup of secrets,
+        or set it through `skip_secrets_setup` in the `.controlplane/controlplane.yml` file
       - Runs a post-creation hook after the app is created if `hooks.post_creation` is specified in the `.controlplane/controlplane.yml` file
       - If the hook exits with a non-zero code, the command will stop executing and also exit with a non-zero code
       - Use `--skip-post-creation-hook` to skip the hook if specified in `controlplane.yml`
     DESC
     VALIDATIONS = %w[config templates].freeze
 
-    def call # rubocop:disable Metrics/MethodLength
+    def call # rubocop:disable Metrics/CyclomaticComplexity, Metrics/MethodLength
       templates = config[:setup_app_templates]
 
       app = cp.fetch_gvc
@@ -32,11 +35,16 @@ module Command
               "or run 'cpl apply-template #{templates.join(' ')} -a #{config.app}'."
       end
 
-      create_secret_and_policy_if_not_exist unless config.options[:skip_secret_access_binding]
+      skip_secrets_setup = config.options[:skip_secret_access_binding] ||
+                           config.options[:skip_secrets_setup] || config.current[:skip_secrets_setup]
 
-      Cpl::Cli.start(["apply-template", *templates, "-a", config.app])
+      create_secret_and_policy_if_not_exist unless skip_secrets_setup
 
-      bind_identity_to_policy unless config.options[:skip_secret_access_binding]
+      args = []
+      args.push("--add-app-identity") unless skip_secrets_setup
+      Cpl::Cli.start(["apply-template", *templates, "-a", config.app, *args])
+
+      bind_identity_to_policy unless skip_secrets_setup
       run_post_creation_hook unless config.options[:skip_post_creation_hook]
     end
 
@@ -89,11 +97,6 @@ module Command
 
     def bind_identity_to_policy
       progress.puts
-
-      if cp.fetch_identity(config.identity).nil?
-        raise "Can't bind identity to policy: identity '#{config.identity}' doesn't exist. " \
-              "Please create it or use `--skip-secret-access-binding` to ignore this message."
-      end
 
       step("Binding identity '#{config.identity}' to policy '#{config.secrets_policy}'") do
         cp.bind_identity_to_policy(config.identity_link, config.secrets_policy)